pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere==5.18.0",
    "fastapi>=0.118.0",
    "fastmcp==2.12.3",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.76",
    "langchain-neo4j==0.5.0",
    "langchain-openai==0.3.33",
    "neomodel>=5.5.3",
    "pandas==2.3.2",
    "pendulum>=3.1.0",
    "pydantic==2.11.9",
    "python-dotenv==1.1.1",
    "pyvi>=0.1.1",
    "rank-bm25==0.2.2",
    "requests==2.32.5",
    "safetensors==0.6.2",
    "sentence-transformers>=5.1.1",
    "tabulate>=0.9.0",
    "torch==2.8.0",
    "transformers==4.56.2",
    "underthesea>=8.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
]
